Jeez.. let's see...
Fixes and changes:
Fixes:
I had a catastrophic issue with the DB which I failed to predict, so I had to resolve that.

Changes:
As heavily advised, I've started using ViewModels in my views, rather than Models on its own... in most cases, there is practically no difference, but it is future proofed now at least. I also implemented some services as 'middleware' between the controllers and the DB context.